Moved helpers to their own solution folder.
