+++ BalancedFieldLength/Application/src/Application.BalancedFieldLength/Views/TabViews/GeneralSImulationSettingsTabView.xaml.cs
@@ -17,6 +17,7 @@
 
 using System.Windows.Controls;
 using WPF.Components;
+using WPF.Components.TextBoxHelpers;
 
 namespace Application.BalancedFieldLength.Views.TabViews
 {
