+++ BalancedFieldLength/WPF/src/WPF.Components/TextBoxHelpers/TextBoxModificationBinding.cs
@@ -16,21 +16,22 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 using System;
-using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Data;
 using System.Windows.Input;
 
-namespace WPF.Components
+namespace WPF.Components.TextBoxHelpers
 {
     /// <summary>
-    /// Helper class which moves the focus of a text box to its next control element after hitting the return key.
+    /// Helper class which commits values of a text box after losing focus or hitting the return key.
     /// </summary>
-    public static class TextBoxMoveFocusOnModification
+    public static class TextBoxModificationBinding
     {
         /// <summary>
-        /// Creates a move focus behaviour on the input argument after the return key was pressed.
+        /// Creates a modification binding on the input argument to commit values after losing
+        /// focus or hitting the return key.
         /// </summary>
-        /// <param name="textBox">The <see cref="TextBox"/> to move the focus for for.</param>
+        /// <param name="textBox">The <see cref="TextBox"/> to create the binding for.</param>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="textBox" />
         /// is <c>null</c>.</exception>
         public static void Create(TextBox textBox)
@@ -40,15 +41,22 @@ namespace WPF.Components
                 throw new ArgumentNullException(nameof(textBox));
             }
 
-            textBox.KeyDown += (sender, e) => OnKeyDown(e, textBox);
+            BindingExpression textBoxBinding = textBox.GetBindingExpression(TextBox.TextProperty);
+            textBox.KeyDown += (sender, e) => OnKeyDown(e, textBoxBinding);
+            textBox.LostFocus += (sender, e) => OnLostFocus(textBoxBinding);
         }
 
-        private static void OnKeyDown(KeyEventArgs e, UIElement textBox)
+        private static void OnKeyDown(KeyEventArgs e, BindingExpressionBase binding)
         {
             if (e.Key.Equals(Key.Return))
             {
-                textBox.MoveFocus(new TraversalRequest(FocusNavigationDirection.Next));
+                binding.UpdateSource();
             }
+        }
+
+        private static void OnLostFocus(BindingExpressionBase binding)
+        {
+            binding.UpdateSource();
         }
     }
 }
