+++ BalancedFieldLength/WPF/src/WPF.Components/WPF.Components.csproj
@@ -33,8 +33,8 @@
       <DependentUpon>TabControlView.xaml</DependentUpon>
     </Compile>
     <Compile Include="TabControl\TabControlViewModel.cs" />
-    <Compile Include="TextBoxMoveFocusOnModification.cs" />
-    <Compile Include="TextBoxModificationBinding.cs" />
+    <Compile Include="TextBoxHelpers\TextBoxMoveFocusOnModification.cs" />
+    <Compile Include="TextBoxHelpers\TextBoxModificationBinding.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\..\licenses\Copying.licenseheader">
